Allow unloading of low priority thumbnails.
This saves a lot of memory but can result in placeholders being seen
during rapid scrolling. Usually only jumps or grabbing the scroll bar.

Use the high priority thumbnail tasks for medium priority thumbnails
when low priority tasks are disabled.

+++ src/config.rs
@@ -128,7 +128,7 @@ pub struct Config {
     #[serde(default)]
     pub max_thumbnailers: u8,
     #[serde(default)]
-    pub background_thumbnailers: u8,
+    pub background_thumbnailers: i16,
     #[serde(default)]
     pub force_small_thumbnails: bool,
 }
